+++ Cargo.toml
@@ -35,3 +35,7 @@ regex = "1"
 
 [profile.dev]
 opt-level = 3
+
+[profile.release]
+opt-level = 3
+lto = true
